Refactor worker completion handling
